INFT-1440

+++ src/Oxa/VideoBundle/Manager/VideoManager.php
@@ -3,7 +3,7 @@
 namespace Oxa\VideoBundle\Manager;
 
 use Domain\SiteBundle\Utils\Helpers\SiteHelper;
-use FFMpeg\Format\Video\WebM;
+use FFMpeg\Format\Video\X264;
 use Gaufrette\Filesystem;
 use Oxa\VideoBundle\Entity\VideoMedia;
 use Symfony\Component\DependencyInjection\ContainerInterface;
@@ -182,8 +182,9 @@ class VideoManager
            if (!file_exists($path)) {
                 mkdir($path);
            }
-
-           $video->save(new WebM(), $path . $name);
+           $format = new X264();
+           $format->setAudioCodec("libmp3lame");
+           $video->save($format, $path . $name);
        } catch (\Exception $e){
            $media->setStatus($media::VIDEO_STATUS_ERROR);
 
@@ -198,12 +199,19 @@ class VideoManager
        ];
        $uploadedFileData = $this->uploadLocalFileData($fileData);
 
+       if ($this->filesystem->getAdapter()->exists($media->getFilepath() . $media->getFilename())) {
+           $this->filesystem->delete($media->getFilepath() . $media->getFilename());
+       }
+
        $media->setName($uploadedFileData['name']);
        $media->setFilepath($uploadedFileData['filepath']);
        $media->setFilename($uploadedFileData['filename']);
        $media->setType($uploadedFileData['type']);
        $media->setStatus($media::VIDEO_STATUS_ACTIVE);
-       unlink($path . $name);
+
+       if (file_exists($path . $name)) {
+           unlink($path . $name);
+       }
 
        return $media;
     }
